First Brute force templating of debris and container objects
