+++ src/satellitePropagator/debris/DebrisContainer.h
@@ -11,12 +11,13 @@ namespace Debris {
  * @class DebrisContainer
  * @brief Holds a vector of Debris objects
  */
+template<class D>
 class DebrisContainer {
 public:
     /**
      * @brief Default constructor
      *
-     * Creates a nwe DebrisContainer object and initializes the #debris_vector
+     * Creates a new DebrisContainer object and initializes the #debris_vector
      * empty
      */
     DebrisContainer();
@@ -52,7 +53,7 @@ public:
      *
      * @param debris Reference to the Debris object to add
      */
-    void addDebris(const Debris& debris);
+    void addDebris(const D& debris);
 
     /**
      * @brief Clear the #debris_vector
@@ -64,7 +65,7 @@ public:
     void cleanDebrisVector();
 
 private:
-    std::vector<Debris> debris_vector; /**< std::vector holding the Debris objects
+    std::vector<D> debris_vector; /**< std::vector holding the Debris objects
                                       this DebrisContainer objects holds */
 public:
     /**
@@ -72,14 +73,49 @@ public:
      *
      * @return Value of #debris_vector
      */
-    [[nodiscard]] const std::vector<Debris>& getDebrisVector() const;
-    std::vector<Debris>& getDebrisVector();
+    [[nodiscard]] const std::vector<D>& getDebrisVector() const;
+    std::vector<D>& getDebrisVector();
 
     /**
      * @brief Setter function for #debris_vector
      *
      * @param debrisVector New value of #debris_vector
      */
-    void setDebrisVector(std::vector<Debris>& debrisVector);
+    void setDebrisVector(std::vector<D>& debrisVector);
 };
+    template<class D>
+    DebrisContainer<D>::DebrisContainer() = default;
+
+    template<class D>
+    DebrisContainer<D>::~DebrisContainer() = default;
+
+    template <class D>
+    void DebrisContainer<D>::addDebris(const D& debris)
+    {
+        debris_vector.push_back(debris);
+    }
+
+    template<class D>
+    void DebrisContainer<D>::cleanDebrisVector()
+    {
+        debris_vector.clear();
+    }
+
+    template <class D>
+    const std::vector<D>& DebrisContainer<D>::getDebrisVector() const
+    {
+        return debris_vector;
+    }
+
+    template <class D>
+    std::vector<D>& DebrisContainer<D>::getDebrisVector()
+    {
+        return debris_vector;
+    }
+
+    template <class D>
+    void DebrisContainer<D>::setDebrisVector(std::vector<D>& debrisVector)
+    {
+        debris_vector = debrisVector;
+    }
 } // namespace Debris
