+++ src/satellitePropagator/io/FileOutput.h
@@ -8,24 +8,29 @@
 #include <iomanip>
 
 #include "satellitePropagator/debris/DebrisContainer.h"
+#include "satellitePropagator/physics/AccelerationComponents/AccelerationComponents.h"
+
+/**
+* @enum OutputFileType
+*
+* @brief Enumerates the possible file types used for output
+*/
+namespace OutputFile {
+    enum Type {
+        TXT, /**< File with .txt extension. Lines are in the format
+    "<token>=<value>" and lines starting with "#" are ignored*/
+        CSV /**< output data into a .csv file */
+    };
+}
+
 /**
  * @class FileOutput
  *
  * @brief Writes output data to a file
  */
+template<class Container, class D>
 class FileOutput {
 public:
-    /**
-     * @enum Type
-     *
-     * @brief Enumerates the possible file types used for output
-     */
-    enum Type {
-        TXT, /**< File with .txt extension. Lines are in the format
-        "<token>=<value>" and lines starting with "#" are ignored*/
-        CSV /**< output data into a .csv file */
-    };
-
     /**
      * @brief Creates new FileOutput object
      *
@@ -36,12 +41,12 @@ public:
      * Debris::Debris objects data from
      * @param output_file_path_arg Path to the main output file to write data
      * to
-     * @param output_file_type_arg FileOutput::Type of the output file
+     * @param output_file_type_arg OutputFileType of the output file
      * @param acc_config Configuration vector for acceleration output
      */
-    FileOutput(Debris::DebrisContainer& debris_arg,
+    FileOutput(Container& debris_arg,
         std::filesystem::path output_file_path_arg,
-        Type output_file_type_arg,
+        OutputFile::Type output_file_type_arg,
         std::array<bool, 8>& acc_config);
 
     /**
@@ -55,7 +60,7 @@ public:
      * @brief Writes data to a file
      *
      * Writes the data to the file with the #output_file_path
-     * by calling a specialized function depending on the FileOutput::Type
+     * by calling a specialized function depending on the OutputFileType
      * #output_file_type
      */
     void writeDebrisData(double t);
@@ -100,11 +105,11 @@ private:
      */
     void writeDebrisCSV(double t);
 
-    Debris::DebrisContainer*
+    Container*
         debris; /**< Reference to a Debris::DebrisContainer object to add
              Debris::Debris objects read from the input file*/
     std::filesystem::path output_file_path; /**< Path to the main output file*/
-    Type output_file_type = CSV; /**< FileOutput::Type of the output file*/
+    OutputFile::Type output_file_type = OutputFile::CSV; /**< OutputFileType of the output file*/
     std::ofstream out; /**< output file stream*/
     double row_count = 0; /**< keeps track of the row number for indexing the lines */
     std::filesystem::path acc_output_file_path; /**< Path to the acceleration output file*/
@@ -116,15 +121,15 @@ public:
      *
      * @return Value of #debris
      */
-    const Debris::DebrisContainer& getDebris() const;
-    Debris::DebrisContainer& getDebris();
+    const Container& getDebris() const;
+    Container& getDebris();
 
     /**
      * @brief Setter function for #debris
      *
      * @param debris New value of #debris
      */
-    void setDebris(Debris::DebrisContainer& debris);
+    void setDebris(Container& debris);
 
     /**
      * @brief Getter function for #output_file_path
@@ -146,14 +151,14 @@ public:
      *
      * @return New value of #output_file_type
      */
-    Type getOutputFileType() const;
+    OutputFile::Type getOutputFileType() const;
 
     /**
      * @brief Setter function for #output_file_type
      *
      * @param outputFileType Value of #output_file_type
      */
-    void setOutputFileType(Type outputFileType);
+    void setOutputFileType(OutputFile::Type outputFileType);
 
     /**
       * @brief Getter function for #acc_output_file_path
@@ -170,3 +175,202 @@ public:
      */
     void setAccOutputFilePath(const std::filesystem::path& accOutputFilePath);
 };
+
+
+template<class Container, class D>
+FileOutput<Container, D>::FileOutput(Container& debris_arg,
+                                     std::filesystem::path output_file_path_arg,
+                                     OutputFile::Type output_file_type_arg,
+                                     std::array<bool, 8>& acc_config)
+        : debris(&debris_arg)
+        , output_file_path(output_file_path_arg)
+        , output_file_type(output_file_type_arg)
+        , out(output_file_path)
+        , acc_output_file_path(output_file_path_arg.replace_extension().concat("_acc.csv"))
+        , acc_out(acc_output_file_path)
+{
+    out << std::setprecision(std::numeric_limits<double>::digits10 + 1);
+    acc_out << std::setprecision(std::numeric_limits<double>::digits10 + 1);
+    if (output_file_type == OutputFile::CSV) {
+        // row counter starts at 0
+        row_count = 0;
+        acc_row_count = 0;
+        // column names
+        out << "index,";
+        out << "time,";
+        out << "position x,";
+        out << "position y,";
+        out << "position z,";
+        out << "position norm,";
+        out << "velocity x,";
+        out << "velocity y,";
+        out << "velocity z,";
+        out << "velocity norm,";
+        out << "acc_t0 x,";
+        out << "acc_t0 y,";
+        out << "acc_t0 z,";
+        out << "acc_t0 norm,";
+        out << "acc_t1 x,";
+        out << "acc_t1 y,";
+        out << "acc_t1 z,";
+        out << "acc_t1 norm" << std::endl;
+
+        // column names for the accelerations
+        acc_out << "index,";
+        acc_out << "time,";
+        if (acc_config[Acceleration::KEP]) {
+            acc_out << "acc_kep x,acc_kep y,acc_kep z,acc_kep norm,";
+        }
+        if (acc_config[Acceleration::J2]) {
+            acc_out << "acc_j2 x,acc_j2 y,acc_j2 z,acc_j2 norm,";
+        }
+        if (acc_config[Acceleration::C22]) {
+            acc_out << "acc_c22 x,acc_c22 y,acc_c22 z,acc_c22 norm,";
+        }
+        if (acc_config[Acceleration::S22]) {
+            acc_out << "acc_s22 x,acc_s22 y,acc_s22 z,acc_s22 norm,";
+        }
+        if (acc_config[Acceleration::SOL]) {
+            acc_out << "acc_sol x,acc_sol y,acc_sol z,acc_sol norm,";
+        }
+        if (acc_config[Acceleration::LUN]) {
+            acc_out << "acc_lun x,acc_lun y,acc_lun z,acc_lun norm,";
+        }
+        if (acc_config[Acceleration::SRP]) {
+            acc_out << "acc_srp x,acc_srp y,acc_srp z,acc_srp norm,";
+        }
+        if (acc_config[Acceleration::DRAG]) {
+            acc_out << "acc_drag x,acc_drag y,acc_drag z,acc_drag norm,";
+        }
+        acc_out << "acc_total x,acc_total y,acc_total z,acc_total norm" << std::endl;
+    }
+}
+
+template<class Container, class D>
+FileOutput<Container, D>::~FileOutput() = default;
+
+template<class Container, class D>
+void FileOutput<Container, D>::writeDebrisData(double t)
+{
+    switch (output_file_type) {
+        case OutputFile::TXT:
+            writeDebrisTXT(t);
+            break;
+        case OutputFile::CSV:
+            writeDebrisCSV(t);
+            break;
+        default:
+            break;
+    }
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::writeDebrisCSV(double t)
+{
+    for (auto d : debris->getDebrisVector()) {
+        out << row_count++ << ',';
+        out << t << ',';
+        IOUtils::to_ostream(d.getPosition(), out);
+        out << ',' << d.getHeight() << ',';
+        IOUtils::to_ostream(d.getVelocity(), out);
+        out << ',' << d.getSpeed() << ',';
+        IOUtils::to_ostream(d.getAccT0(), out);
+        out << ',' << d.getAccT0Norm() << ',';
+        IOUtils::to_ostream(d.getAccT1(), out);
+        out << ',' << d.getAccT1Norm();
+        out << '\n';
+    }
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::writeDebrisTXT(double t)
+{
+    // NOT IMPLEMENTED YET
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::writeAcc_start(double t)
+{
+    acc_out << acc_row_count++ << ',';
+    acc_out << t << ',';
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::writeAcc_value(const std::array<double, 3>& vec)
+{
+    IOUtils::to_ostream(vec, acc_out);
+    acc_out << "," << MathUtils::euclideanNorm(vec) << ",";
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::writeAcc_end(const std::array<double, 3>& vec)
+{
+    IOUtils::to_ostream(vec, acc_out);
+    acc_out << "," << MathUtils::euclideanNorm(vec) << "\n";
+}
+
+template<class Container, class D>
+const Container& FileOutput<Container, D>::getDebris() const
+{
+    return *debris;
+}
+
+template<class Container, class D>
+Container& FileOutput<Container, D>::getDebris()
+{
+    return *debris;
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::setDebris(Container& debris)
+{
+    FileOutput<Container, D>::debris = &debris;
+}
+
+template<class Container, class D>
+const std::filesystem::path& FileOutput<Container, D>::getOutputFilePath() const
+{
+    return output_file_path;
+}
+
+template<class Container, class D>
+std::filesystem::path& FileOutput<Container, D>::getOutputFilePath()
+{
+    return output_file_path;
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::setOutputFilePath(const std::filesystem::path& outputFilePath)
+{
+    output_file_path = outputFilePath;
+}
+
+template<class Container, class D>
+OutputFile::Type FileOutput<Container, D>::getOutputFileType() const
+{
+    return output_file_type;
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::setOutputFileType(OutputFile::Type outputFileType)
+{
+    output_file_type = outputFileType;
+}
+
+template<class Container, class D>
+const std::filesystem::path& FileOutput<Container, D>::getAccOutputFilePath() const
+{
+    return acc_output_file_path;
+}
+
+template<class Container, class D>
+std::filesystem::path& FileOutput<Container, D>::getAccOutputFilePath()
+{
+    return acc_output_file_path;
+}
+
+template<class Container, class D>
+void FileOutput<Container, D>::setAccOutputFilePath(const std::filesystem::path& accOutputFilePath)
+{
+    acc_output_file_path = accOutputFilePath;
+}
