+++ src/satellitePropagator/physics/AccelerationComponents/DragComponent.h
@@ -19,7 +19,35 @@ namespace {
      * @param acc_total Reference to an 3D vector to accumulate the accelerations
      * for all applied Acceleration::AccelerationComponent.
      */
-void apply(const Debris::Debris& d,
+template<class D>
+void
+apply(const D& d,
     std::array<double, 3>& acc_drag,
     std::array<double, 3>& acc_total);
+
+
+    namespace {
+    }
+    template<class D>
+    void
+    apply(const D& d,
+          std::array<double, 3>& acc_drag,
+          std::array<double, 3>& acc_total)
+    {
+
+        // calculate atmospheric density
+        const double p = Physics::P_GROUND * std::exp(-(d.getHeight() - Physics::R_EARTH) / Physics::H_ATMOSPHERE);
+        const double f = -0.5 * p * d.getBcInv();
+        // calculate velocity relative to atmosphere
+        acc_drag = d.getVelocity();
+        acc_drag[0] += Physics::ROT_ATMOSPHERE * d.getPosition()[1];
+        acc_drag[1] -= Physics::ROT_ATMOSPHERE * d.getPosition()[0];
+        // actual drag
+        acc_drag[0] = f * acc_drag[0] * acc_drag[0];
+        acc_drag[1] = f * acc_drag[1] * acc_drag[1];
+        acc_drag[2] = f * acc_drag[2] * acc_drag[2];
+        acc_total[0] += acc_drag[0];
+        acc_total[1] += acc_drag[1];
+        acc_total[2] += acc_drag[2];
+    }
 } // namespace Acceleration
