+++ src/satellitePropagator/physics/AccelerationComponents/SRPComponent.h
@@ -14,7 +14,10 @@ namespace {
          *
          * @return Physics::
          */
-    inline constexpr double getFactor();
+    inline constexpr double getFactor()
+    {
+        return Physics::P_SRP * Physics::AU * Physics::AU;
+    }
 }
 /**
      * @brief Precalculates values needed for Acceleration::SolComponent::apply()
@@ -39,9 +42,37 @@ std::array<double, 6> setUp(double t);
      * @param acc_total Reference to an 3D vector to accumulate the accelerations
      * for all applied Acceleration::AccelerationComponent.
      */
-void apply(const Debris::Debris& d,
+template<class D>
+void
+apply(const D& d,
     double d_srp,
     const std::array<double, 6>& sun_params,
     std::array<double, 3>& acc_srp,
     std::array<double, 3>& acc_total);
+
+    template<class D>
+    void
+    apply(const D& d,
+          double d_srp,
+          const std::array<double, 6>& sun_params,
+          std::array<double, 3>& acc_srp,
+          std::array<double, 3>& acc_total)
+    {
+        acc_srp = d.getPosition();
+        acc_srp[0] -= sun_params[0];
+        acc_srp[1] -= sun_params[1];
+        acc_srp[2] -= sun_params[2];
+        // check if we have a precalculated value for d_srp and calculate it if needed
+        if (d_srp == 0.0) {
+            d_srp = std::inner_product(acc_srp.begin(), acc_srp.end(), acc_srp.begin(), .0);
+            d_srp = 1 / std::sqrt(d_srp * d_srp * d_srp);
+        }
+        // Eq 50
+        acc_srp[0] = d.getAom() * (getFactor() * d_srp * acc_srp[0]);
+        acc_srp[1] = d.getAom() * (getFactor() * d_srp * acc_srp[1]);
+        acc_srp[2] = d.getAom() * (getFactor() * d_srp * acc_srp[2]);
+        acc_total[0] += acc_srp[0];
+        acc_total[1] += acc_srp[1];
+        acc_total[2] += acc_srp[2];
+    }
 } // namespace Acceleration
