+++ test/physics_test/Acceleration_test.h
@@ -13,12 +13,12 @@
 
 class KepComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
         // only one axis
         for (int i = 0; i < 3; ++i) {
@@ -69,8 +69,8 @@ protected:
             -(1 / (std::sqrt(3) * 3000000)) * Physics::GM_EARTH,
             -(1 / (std::sqrt(3) * 3000000)) * Physics::GM_EARTH };
     }
-
-    void calcKep(Debris::Debris& d, std::array<double, 3>& acc_kep)
+    template<class D>
+    void calcKep(D& d, std::array<double, 3>& acc_kep)
     {
         double gme = 3.986004407799724e+5;
         double x = d.getPosition()[0];
@@ -85,12 +85,12 @@ protected:
 
 class J2ComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
 
         for (int i = 0; i < 3; ++i) {
@@ -134,7 +134,8 @@ protected:
         };
     }
 
-    void calcJ2(Debris::Debris& d, std::array<double, 3>& acc_j2)
+    template<class D>
+    void calcJ2(D& d, std::array<double, 3>& acc_j2)
     {
         double gme = 3.986004407799724e+5;
         double re = 6378.1363;
@@ -156,12 +157,12 @@ protected:
 
 class C22ComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
 
         for (int i = 0; i < 3; ++i) {
@@ -194,7 +195,8 @@ protected:
         pre_calculated[8] = { 0, 0, 0 };
     }
 
-    void calcC22(Debris::Debris& d, double t, std::array<double, 3>& acc_c22)
+    template<class D>
+    void calcC22(D& d, double t, std::array<double, 3>& acc_c22)
     {
         double gme = 3.986004407799724e+5;
         double re = 6378.1363;
@@ -226,12 +228,12 @@ protected:
 
 class S22ComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
 
         for (int i = 0; i < 3; ++i) {
@@ -263,7 +265,8 @@ protected:
         pre_calculated[8] = { 0, 0, 0 };
     }
 
-    void calcS22(Debris::Debris& d, double t, std::array<double, 3>& acc_s22)
+    template<class D>
+    void calcS22(D& d, double t, std::array<double, 3>& acc_s22)
     {
         double pi = std::acos(-1);
         double gme = 3.986004407799724e+5;
@@ -296,12 +299,12 @@ protected:
 
 class C22S22ComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
 
         for (int i = 0; i < 3; ++i) {
@@ -325,12 +328,12 @@ protected:
 
 class LunComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
 
         for (int i = 0; i < 3; ++i) {
@@ -456,7 +459,8 @@ protected:
         return moon_params;
     }
 
-    void calcLun(Debris::Debris& d, double t, std::array<double, 3>& acc_lun)
+    template<class D>
+    void calcLun(D& d, double t, std::array<double, 3>& acc_lun)
     {
         double gmm = 4.9028e+3;
         double e = 23.4392911;
@@ -547,12 +551,12 @@ protected:
 
 class SolComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
         double t = 0;
 
@@ -576,7 +580,8 @@ protected:
         pre_calculated[8] = { 0, 0, 0 };
     }
 
-    void calcSol(Debris::Debris& d, double t, std::array<double, 3>& acc_sol)
+    template<class D>
+    void calcSol(D& d, double t, std::array<double, 3>& acc_sol)
     {
         double gms = 1.32712440018e+11;
         double phis0 = 357.5256;
@@ -638,12 +643,12 @@ protected:
 
 class SRPComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
         double t = 0;
 
@@ -668,7 +673,8 @@ protected:
         pre_calculated[8] = { 0, 0, 0 };
     }
 
-    void calcSRP(Debris::Debris& d, double t, std::array<double, 3>& acc_srp)
+    template<class D>
+    void calcSRP(D& d, double t, std::array<double, 3>& acc_srp)
     {
         double p = 4.56e-3;
         double a = 1.49619e+8;
@@ -707,12 +713,12 @@ protected:
 
 class DragComponentTests : public ::testing::Test {
 protected:
-    std::shared_ptr<Debris::DebrisContainer> debris;
+    std::shared_ptr<Debris::DebrisContainer<Debris::Debris>> debris;
     std::array<std::array<double, 3>, 9> pre_calculated;
 
     virtual void SetUp()
     {
-        debris = std::make_shared<Debris::DebrisContainer>();
+        debris = std::make_shared<Debris::DebrisContainer<Debris::Debris>>();
         Debris::Debris d;
 
         for (int i = 0; i < 3; ++i) {
@@ -734,7 +740,8 @@ protected:
             debris->addDebris(d);
         }
     }
-    void calcDrag(Debris::Debris& d, std::array<double, 3>& acc_drag)
+    template<class D>
+    void calcDrag(D& d, std::array<double, 3>& acc_drag)
     {
         double re = 6378.1363;
         double p0 = 1.3;
